+++ src/index.ts
@@ -1,17 +1,21 @@
-import { validateSkippedStacks } from '@mnrendra/read-stacked-json'
-
 import mainAsync from './async'
 import mainSync from './sync'
 
 export type {
-  Package,
-  Options,
-  SkippedStacks,
-  ValidSkippedStacks
+  Package
+} from '@mnrendra/types-package'
+
+export {
+  type SkippedStacks,
+  type ValidSkippedStacks,
+  validateSkippedStacks
+} from '@mnrendra/read-stacked-json'
+
+export type {
+  Options
 } from './types'
 
 export {
   mainAsync as readPackage,
-  mainSync as readPackageSync,
-  validateSkippedStacks
+  mainSync as readPackageSync
 }
