+++ src/types/index.ts
@@ -1,14 +1,5 @@
 import type Options from './Options'
 
 export type {
-  Package
-} from '@mnrendra/types-package'
-
-export type {
-  SkippedStacks,
-  ValidSkippedStacks
-} from '@mnrendra/read-stacked-json'
-
-export type {
   Options
 }
