--- tests/mocks/index.ts
@@ -1,9 +1,0 @@
-import readAsync from './readAsync'
-import readSync from './readSync'
-import stackTrace from './stackTrace'
-
-export {
-  readAsync,
-  readSync,
-  stackTrace
-}
